+++ src/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.projectlearn.login.Main
+Main-Class: com.projectlearn.Main
