--- src/com/projectlearn/finance/Permissions.java
@@ -1,11 +1,0 @@
-package com.projectlearn.finance;
-
-public enum Permissions {
-    CAN_HIRE,
-    CAN_FIRE,
-    CAN_DEPOSIT,
-    CAN_WITHDRAW,
-    CAN_PAY,
-    CAN_FINE,
-    CAN_VIEW_ACCOUNTS;
-}
