--- src/com/projectlearn/finance/finance.css
@@ -1,31 +1,0 @@
-#deposit {
--fx-font-size: 12px ;
--fx-background-color: #dcdcdc;
--fx-padding: 4 4 4 4;
--fx-text-fill: #800000;
--fx-border-radius: 8;
--fx-font-weight: bold;
-}
-
-#withdraw {
--fx-font-size: 12px ;
--fx-background-color: #dcdcdc;
--fx-padding: 4 4 4 4;
--fx-text-fill: #800000;
--fx-border-radius: 8;
--fx-font-weight: bold;
-}
-
-#warning {
--fx-font-size: 13px;
--fx-text-fill: #800000;
--fx-font-weight: bold;
-}
-
-.info-labels {
--fx-text-fill: #0000cd;
-}
-
-.root {
--fx-background-color: #f5f5dc
-}
