+++ src/com/projectlearn/finance/gui/accountsView.fxml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+
+<AnchorPane prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.projectlearn.finance.gui.AccountViewController">
+   <children>
+      <TableView fx:id="accounts" layoutX="30.0" layoutY="100.0" prefHeight="400.0" prefWidth="743.0" AnchorPane.bottomAnchor="100.0" AnchorPane.leftAnchor="30.0" AnchorPane.rightAnchor="27.0" AnchorPane.topAnchor="100.0">
+        <columns>
+          <TableColumn fx:id="accountNums" prefWidth="120.0" text="Account Number" />
+          <TableColumn fx:id="accountBalance" prefWidth="75.0" text="Balance" />
+            <TableColumn fx:id="name" prefWidth="124.0" text="Name" />
+            <TableColumn fx:id="email" prefWidth="235.0" text="Email" />
+            <TableColumn fx:id="id" minWidth="0.0" prefWidth="71.0" text="ID" />
+            <TableColumn fx:id="type" prefWidth="114.0" text="Type" />
+        </columns>
+      </TableView>
+      <HBox layoutX="54.0" layoutY="539.0" prefHeight="32.0" prefWidth="692.0" spacing="20.0" AnchorPane.bottomAnchor="29.0" AnchorPane.leftAnchor="54.0" AnchorPane.rightAnchor="54.0" AnchorPane.topAnchor="539.0">
+         <children>
+            <Button fx:id="back" mnemonicParsing="false" onAction="#back" text="Back" />
+            <Button fx:id="edit" mnemonicParsing="false" onAction="#editAccount" text="Edit" />
+            <Button fx:id="newAccount" mnemonicParsing="false" onAction="#newAccount" text="New" />
+            <Button fx:id="delete" mnemonicParsing="false" onAction="#delete" text="Delete" />
+            <Button fx:id="done" mnemonicParsing="false" onAction="#doneEdit" text="Done" />
+         </children>
+      </HBox>
+   </children>
+</AnchorPane>
