+++ src/com/projectlearn/finance/logic/Permissions.java
@@ -1,0 +1,11 @@
+package com.projectlearn.finance.logic;
+
+public enum Permissions {
+    CAN_HIRE,
+    CAN_FIRE,
+    CAN_DEPOSIT,
+    CAN_WITHDRAW,
+    CAN_PAY,
+    CAN_FINE,
+    CAN_VIEW_ACCOUNTS;
+}
