--- src/com/projectlearn/finance/newAccount.fxml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ChoiceBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="539.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.projectlearn.finance.NewAccountController">
-   <children>
-      <VBox prefHeight="539.0" prefWidth="120.0" spacing="20.0">
-         <children>
-            <Label text="Name" />
-            <TextField fx:id="name" promptText="ex: John Doe" />
-            <Label text="Password" />
-            <PasswordField fx:id="password" promptText="ex: 1234" />
-            <Label text="Email" />
-            <TextField fx:id="email" promptText="ex: johnDoe@doe.com" />
-            <Label text="ID" />
-            <TextField fx:id="ID" promptText="ex: 1" />
-            <Label text="Type" />
-            <ChoiceBox fx:id="type" prefWidth="150.0" />
-         </children>
-         <padding>
-            <Insets left="10.0" />
-         </padding>
-      </VBox>
-      <Label fx:id="alreadyInUse" layoutX="130.0" layoutY="14.0" prefHeight="50.0" prefWidth="665.0" textFill="DARKRED">
-         <font>
-            <Font size="18.0" />
-         </font>
-      </Label>
-      <HBox layoutX="586.0" layoutY="500.0" prefHeight="25.0" prefWidth="135.0" spacing="50.0">
-         <children>
-            <Button fx:id="quit" mnemonicParsing="false" onAction="#quitCreation" text="Quit" />
-            <Button fx:id="done" mnemonicParsing="false" onAction="#addNewAccount" text="Done" />
-         </children>
-      </HBox>
-   </children>
-</AnchorPane>
