+++ src/com/projectlearn/login/gui/login.css
@@ -1,0 +1,3 @@
+.root {
+-fx-background-color: #f5f5dc
+}
