--- src/com/projectlearn/login/login.css
@@ -1,3 +1,0 @@
-.root {
--fx-background-color: #f5f5dc
-}
