--- src/com/projectlearn/login/login.fxml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.Region?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="545.0" prefWidth="364.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.projectlearn.login.LoginController">
-   <children>
-      <VBox layoutX="21.0" layoutY="244.0" prefHeight="156.0" prefWidth="324.0">
-         <children>
-            <Label fx:id="accountLabel" prefHeight="28.0" prefWidth="137.0" text="Account Number">
-               <font>
-                  <Font size="18.0" />
-               </font>
-            </Label>
-            <TextField fx:id="numTextField" prefHeight="33.0" prefWidth="311.0" promptText="Enter your email">
-               <font>
-                  <Font size="18.0" />
-               </font>
-            </TextField>
-            <VBox prefHeight="79.0" prefWidth="324.0">
-               <children>
-                  <Label fx:id="passwordLabel" prefHeight="28.0" prefWidth="92.0" text="Password">
-                     <font>
-                        <Font size="18.0" />
-                     </font>
-                  </Label>
-                  <PasswordField fx:id="passwordField" prefHeight="33.0" prefWidth="311.0" promptText="Enter your password">
-                     <font>
-                        <Font size="18.0" />
-                     </font>
-                  </PasswordField>
-               </children>
-            </VBox>
-         </children>
-      </VBox>
-      <HBox layoutX="127.0" layoutY="460.0" prefHeight="39.0" prefWidth="217.0" AnchorPane.bottomAnchor="46.0" AnchorPane.leftAnchor="127.0" AnchorPane.rightAnchor="20.0" AnchorPane.topAnchor="460.0">
-         <children>
-            <Button fx:id="cancelBtn" contentDisplay="RIGHT" mnemonicParsing="false" onAction="#cancelBtn_Click" prefHeight="39.0" prefWidth="99.0" text="Cancel" textAlignment="CENTER">
-               <font>
-                  <Font size="18.0" />
-               </font>
-            </Button>
-            <Region prefHeight="53.0" prefWidth="18.0" scaleZ="-14.0" />
-            <Button fx:id="loginBtn" contentDisplay="RIGHT" mnemonicParsing="false" onAction="#loginBtn_Click" prefHeight="39.0" prefWidth="99.0" text="Login" textAlignment="CENTER">
-               <font>
-                  <Font size="18.0" />
-               </font>
-            </Button>
-         </children>
-      </HBox>
-      <Text fx:id="loginText" fill="#42c90d" layoutY="176.0" strokeType="OUTSIDE" strokeWidth="0.0" textAlignment="CENTER" wrappingWidth="364.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0">
-         <font>
-            <Font size="28.0" />
-         </font>
-      </Text>
-      <Button fx:id="newAccountBtn" layoutX="257.0" layoutY="506.0" mnemonicParsing="false" onAction="#newAccount" text="New Account" />
-      <ImageView fx:id="logoImage" fitHeight="156.0" fitWidth="305.0" layoutX="30.0" layoutY="14.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../finance/ProjectLearn_Logo.png" />
-         </image></ImageView>
-   </children>
-</AnchorPane>
